% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoBreaks.R
\name{AutoBreaks}
\alias{AutoBreaks}
\title{Automatically Create Labelled Breaks for Continuous Data}
\usage{
AutoBreaks(x, by, ...)
}
\arguments{
\item{x}{input numeric or integer vector}

\item{by}{numeric breaks to be applied. Can take three types of breaks: i) Fixed, ii) Defined, or iii) Split. See 'Details'.}

\item{...}{Set decimal place for integers. Default is 2. Can provide value only. Optional.}
}
\value{
Returns a character vector of the same length as input. Missing values written as NA.
}
\description{
\code{AutoBreaks} returns a character vector of the same length as input x. Each value of x is replaced by a label matching the break parameters. 
                  Break parameters can either be fixed, defined, or split. See details.
}
\details{
The values for paremeter \code{by} can be supplied in 1 of 3 ways:
\itemize{
\item Fixed Breaks: supply a single value. This feature automatically sets break points by value suppled starting from the minimum value.
\item Defined Breaks: supply a sequnce of values in ascending order. User defines break points and the number of categories.
\item Split Breaks: supply two of same values. This feature will automatically split all values into one of the two categories.
}
}
\examples{
# Example Data
norm.int <- runif(1000, 0, 1)
cont.int <- runif(1000, 1, 15)
rand.num <- sample(20:84, 100, replace = T)

# Fixed breaks (takes a single value):
AutoBreaks(rand.num, by = 10)
AutoBreaks(cont.int, 3)
AutoBreaks(norm.int, 0.2)

# Defined breaks (user-defined breaks):
# NB: Starting value can be 0 if required
AutoBreaks(rand.num, by = c(30,45,65))
AutoBreaks(cont.int, c(2,3,4,5))
AutoBreaks(norm.int, c(0,0.8))

# Split break
AutoBreaks(rand.num, c(55, 55))
AutoBreaks(norm.int, by = c(0.8, 0.8), 6) # can set decimal place

# Can set decimal place
AutoBreaks(cont.int, c(3,4,5), dp=4)
AutoBreaks(norm.int, c(0.4,0.6,0.8), 5)

}
\seealso{
\code{\link{FreqTable}} utilised this function when users attempt to create a frequency table using continous variables.
}
