% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateVectoriser.R
\name{DateVectoriser}
\alias{DateVectoriser}
\title{Date Vectoriser for GroupIntervalDates}
\usage{
DateVectoriser(dat, start, end, lag, zero.index)
}
\arguments{
\item{dat}{a data.table containing input data (see details)}

\item{start}{date variable in the dataset representing the start of an observation}

\item{end}{date variable in the dataset representing the end of an observation}

\item{lag}{specify the lag days; default is 1}

\item{zero.index}{specify whether the indexing sequence should start at 0 instead of 1; default is FALSE}
}
\value{
Returns a sequential vector of the same length as input x
}
\description{
\code{DateVectoriser} returns a sequential vector of binary values in the same length as input x. 
Set first element always the starting position. In subsequent elements; a new index is formed where index position equates 1 and the prior index is 0.
}
