% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqTable.R
\name{FreqTable}
\alias{FreqTable}
\title{Frequency Table for Multiple Variables}
\usage{
FreqTable(data, xvar, yvar, xlab, ylab, margin, pcn, yord, copy)
}
\arguments{
\item{data}{input data as data.frame or data table}

\item{xvar}{single or multiple variables containing comparison categories. See 'Details'.}

\item{yvar}{a single variable containing reference categories}

\item{xlab}{User defined label(s) in the order of the comparison variables used in xvar. Optional. See 'Details'.}

\item{ylab}{User defined label for the reference variable used in yvar. Optional.}

\item{margin}{If TRUE, totals for each group are shown in the table's margin.}

\item{pcn}{Show two types of percentages: "v" or "h". Optional. See 'Details'.}

\item{yord}{User defined ordering for cateogries in the reference variable. Optional. See 'Details'.}

\item{copy}{If TRUE, copies output to clipboard - enabling pasting to excel (using 'Ctrl + v'). Optional.}
}
\value{
Returns a data.frame with a contingency table structure. Missing values written as NA. Can be copied to excel using the argument \code{copy=TRUE}
}
\description{
\code{FreqTable} compiles a contingency table that allows for multiple comparison variables, against a reference variable. 
 Each variable supplied can contain multiple categories. If one or more continuous variables are supplied, users are prompted to label values using \code{AutoBreaks}.
 Users can relabel variable names, add percentages, add a totals margin, and change the order of reference variable categories. Users can easily copy the output contingency table to a spreadsheet.
}
\details{
\itemize{
\item \code{xvar} can supple one or more variables. If multiple are used, then use format \code{c("var1", "var2", "var3")}. 
\item \code{xlab} number of labels supplied must match length and order of xvar. If multiple are used, then use format \code{c("name1", "name2", "name3")}.
\item \code{yord} number of labels supplied must number of levels in yvar. Missing can be ignored.
\item \code{pcn}  use "v" for vertical proportion which shows the \% of total across comparison group(s); else "h" for horizontal proportions which shows the \% of total reference groups. Default is NULL.
\item If numeric or integer variables are supplied, then user will be prompt with \code{AutoBreaks} which automatically labels a numeric data as character strings.
}
}
\examples{
FreqTable(data = SAMPLE, 
          xvar = c("en_nzdep_q", "en_eth","hx_broad_cvd"), 
          yvar = "gender",
          xlab = c("DEPRIVATION", "ETHNICITY", "HISTORY CVD"),
          ylab = "SEX",
          yord = c("M", "F"),
          margin = T,
          pcn = "v",
          copy = T)

FreqTable(data = SAMPLE,
          xvar = c("index_age", "cvd_policyrisk5y_score"),
          yvar = "gender",
          pcn ="h")
          
}
