% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamCheck.R
\name{ParamCheck}
\alias{ParamCheck}
\title{General Parameter Checking for PredictRiskScores}
\usage{
ParamCheck(input, vars, call, is.table)
}
\arguments{
\item{input}{list of all input values inherited from the function call}

\item{vars}{character vector of required parameters}

\item{call}{name of the function being called}

\item{is.table}{logical; whether a dataset is provided in the parent.frame}
}
\value{
\code{ParamCheck} returns warning messages if any invalid values are detected. In cases where input data is invalid, \code{ParamCheck} will
stop the function's operation and return an error message.
}
\description{
\code{ParamCheck} evaluates input parameters inherited from the parent.frame; to check for names, values, and validity and ensure that a risk calculation can be carried out.
The call is stopped completely when required arguments are not stated or when compulsary inputs are invalid.
Invalid values may still be calculated an \code{NA} returned with an error message indicating one or more problematic variables.
}
\details{
\code{ParamCheck} identifies the following errors and invalid inputs:
\enumerate{
 \item missing arguments - operations stops
 \item missing input data - operation stops
 }
}
\author{
Billy Wu (R Developer)
}
